Revert "Merge branch 'FE' into 'master'"
This reverts merge request !30

+++ gaebap/src/main/java/site/doggyyummy/gaebap/domain/meeting/dto/CreateMeetingResponseDTO.java
@@ -18,7 +18,7 @@ public class CreateMeetingResponseDTO {
     private String password;
 
     @JsonProperty("max_participant")
-    private int maxParticipant;
+    private Long maxParticipant;
 
     private String title;
