+++ gaebap/src/main/java/site/doggyyummy/gaebap/domain/meeting/dto/ModifyMeetingResponseDTO.java
@@ -18,7 +18,7 @@ public class ModifyMeetingResponseDTO {
     private String password;
 
     @JsonProperty("max_participant")
-    private int maxParticipant;
+    private Long maxParticipant;
 
     private String title;
